Add XML namespaces to OpdsEngine class and remove OpdsNamespaces class
